+++ src/ResourceManager/Network/Commands.Network/VirtualNetwork/NewAzureVirtualNetworkCommand.cs
@@ -76,12 +76,17 @@ namespace Microsoft.Azure.Commands.Network
             ValueFromPipelineByPropertyName = true,
             HelpMessage = "A hashtable which represents resource tags.")]
         public Hashtable Tag { get; set; }
-        
+
+        [Parameter(
+            Mandatory = false,
+            HelpMessage = "A switch parameter which represents whether DDoS protection is enabled or not. It can only be turned on if a DDoS Protection Plan is associated with the virtual network.")]
+        public SwitchParameter EnableDdosProtection { get; set; }
+
         [Parameter(
             Mandatory = false,
             ValueFromPipelineByPropertyName = true,
-            HelpMessage = "A switch parameter which represents if DDoS protection is enabled or not.")]
-        public SwitchParameter EnableDDoSProtection { get; set; }
+            HelpMessage = "Reference to the DDoS protection plan resource associated with the virtual network.")]
+        public string DdosProtectionPlanId { get; set; }
 
         [Parameter(
            Mandatory = false,
@@ -131,8 +136,14 @@ namespace Microsoft.Azure.Commands.Network
             }
 
             vnet.Subnets = this.Subnet;
-            vnet.EnableDDoSProtection = this.EnableDDoSProtection;
+            vnet.EnableDdosProtection = this.EnableDdosProtection;
             vnet.EnableVmProtection = this.EnableVmProtection;
+
+            if (!string.IsNullOrEmpty(this.DdosProtectionPlanId))
+            {
+                vnet.DdosProtectionPlan = new PSResourceId();
+                vnet.DdosProtectionPlan.Id = this.DdosProtectionPlanId;
+            }
 
             // Map to the sdk object
             var vnetModel = NetworkResourceManagerProfile.Mapper.Map<MNM.VirtualNetwork>(vnet);